Add callbacks for adding / removing child Node
This commits adds a callback mechanism when a Node's
child gets added / removed.

Because it would IMHO be costly to have callbacks registered for each Node,
I choose to put the callbacks list in the Device.

If a node is inside the hierarchy of a device (i.e. the important case)
calling emplaceAndNotify / eraseAndNotify will then run the callbacks
registered in the device.

This way a single function has to be registered.

This is used in i-score for the Local device (see recent commits).
This allows to automatically update the Device explorer when for instant
a Constraint is added (instead of having to do Refresh namespace each time).

I kept separate emplace / emplaceAndNotify because if a Node is added
by hand from i-score, we do not want (I think) to get notified of the
node that was just added.