fix weird case where avahi is not found correctly
when crosscompiling, this happen when `avahi-client/avahi.h` is the `${CMAKE_SYSROOT]/usr/include/` folder.
And that default folder is not added automatically to headers search path.
This might be related to https://github.com/abhiTronix/raspberry-pi-cross-compilers/issues/3#issuecomment-453117354.
Btw I have to install avahi libs in /opt/avahi and add this folder to `CMAKE_PREFIX_PATH`.